Update README.md to indicate Go 1.7+ is required to use this package
New functionality added to this package now depends on the `context` package
that was introduced in Go 1.7. This change updates the README to reflect that
new version constraint.

If this becomes an issue, we could add the `go-constraint` package as a
dependency to this. This would allow us to have a compile time check of the Go
version.