Add Windows CI with AppVeyor; fix bugs discovered through Windows CI
This change introduces Continuous Integration of changes in a Windows
environment, using AppVeyor.

In the process of setting up CI a few bugs were discovered in the Windows
implementation, specifically around the behavior of the TryLock() function.

The API for TryLock() was not supposed to return an error when we failed to lock
due to someone else having a lock, because that was an expected failure mode.
However, in the Windows implementation we weren't doing this properly for a few
reasons.

The first reason was that the logic simply wasn't behaving properly in that
situation, and was returning the error out to the consumer of the API. Upon the
fixing of the logic, it was discovered that the code wasn't checking for the
right error number to know when we failed to lock the file due to a preexisting
lock.

Looking at the documentation, there is some conflicting information about what
error number would be returned. Initially this caught the `ERROR_IO_PENDING`
only, and that didn't work in the CI environment. So in addition to that, this
now looks for the error number to be `ERROR_LOCK_VIOLATION`. If either is found,
it returns an indicator that the lock wasn't made but that there wasn't an
error.

Signed-off-by: Tim Heckman <t@heckman.io>