Add comment explaining sharp edges on UNIX (when using shared locks)
This adds comments to the `flock` package around the sharp edges on UNIX-like
operating systems, where the shared file lock can be converted to an exclusive
lock unbeknownst to a caller who has a shared lock. If `Unlock` is called, the
exclusive lock is dropped, which may result in undesired behaviors.

Signed-off-by: Tim Heckman <t@heckman.io>